move update checking into separate component
