+++ resources/js/beatmap-discussions/main.tsx
@@ -5,7 +5,6 @@ import NewReview from 'beatmap-discussions/new-review';
 import { ReviewEditorConfigContext } from 'beatmap-discussions/review-editor-config-context';
 import BackToTop from 'components/back-to-top';
 import BeatmapsetWithDiscussionsJson from 'interfaces/beatmapset-with-discussions-json';
-import { route } from 'laroute';
 import { action, makeObservable, observable, reaction, toJS } from 'mobx';
 import { observer } from 'mobx-react';
 import core from 'osu-core-singleton';
@@ -21,19 +20,14 @@ import { Events } from './events';
 import { Header } from './header';
 import { ModeSwitcher } from './mode-switcher';
 import { NewDiscussion } from './new-discussion';
+import { Refresh } from './refresh';
 
-const checkNewTimeoutDefault = 10000;
-const checkNewTimeoutMax = 60000;
 const beatmapsetJsonId = 'json-beatmapset';
 
 interface Props {
   reviewsConfig: {
     max_blocks: number;
   };
-}
-
-interface UpdateResponseJson {
-  beatmapset: BeatmapsetWithDiscussionsJson;
 }
 
 @observer
@@ -45,10 +39,7 @@ export default class Main extends React.Component<Props> {
   private readonly focusNewDiscussion = currentUrl().hash === '#new';
   private readonly modeSwitcherRef = React.createRef<HTMLDivElement>();
   private readonly newDiscussionRef = React.createRef<HTMLDivElement>();
-  private nextTimeout = checkNewTimeoutDefault;
   @observable private readonly store;
-  private timeoutCheckNew?: number;
-  private xhrCheckNew?: JQuery.jqXHR<UpdateResponseJson>;
 
   constructor(props: Props) {
     super(props);
@@ -83,8 +74,6 @@ export default class Main extends React.Component<Props> {
         }),
       );
     })));
-
-    this.timeoutCheckNew = window.setTimeout(this.checkNew, checkNewTimeoutDefault);
   }
 
   componentWillUnmount() {
@@ -95,6 +84,7 @@ export default class Main extends React.Component<Props> {
   render() {
     return (
       <>
+        <Refresh discussionsState={this.discussionsState} />
         <Header
           discussionsState={this.discussionsState}
           store={this.store}
@@ -140,39 +130,10 @@ export default class Main extends React.Component<Props> {
     );
   }
 
-  @action
-  private readonly checkNew = () => {
-    if (this.xhrCheckNew != null) return;
-
-    window.clearTimeout(this.timeoutCheckNew);
-
-    this.xhrCheckNew = $.get(route('beatmapsets.discussion', { beatmapset: this.discussionsState.beatmapset.id }), {
-      format: 'json',
-      last_updated: this.discussionsState.lastUpdate,
-    });
-
-    this.xhrCheckNew.done((data, _textStatus, xhr) => {
-      if (xhr.status === 304) {
-        this.nextTimeout *= 2;
-        return;
-      }
-
-      this.nextTimeout = checkNewTimeoutDefault;
-      this.discussionsState.update({ beatmapset: data.beatmapset });
-    }).always(() => {
-      this.nextTimeout = Math.min(this.nextTimeout, checkNewTimeoutMax);
-
-      this.timeoutCheckNew = window.setTimeout(this.checkNew, this.nextTimeout);
-      this.xhrCheckNew = undefined;
-    });
-  };
-
   private readonly destroy = () => {
     document.removeEventListener('turbolinks:before-cache', this.destroy);
 
     document.documentElement.style.removeProperty('--scroll-padding-top-extra');
-    window.clearTimeout(this.timeoutCheckNew);
-    this.xhrCheckNew?.abort();
 
     storeJson(beatmapsetJsonId, toJS(this.store.beatmapset));
     this.discussionsState.saveState();
